drm/i915/bios: shrink all Snps's VS/PE tables
Each value present in Snps's VS/PE-O table is stored on u32. In order to
match values' size this causes dg2_snps_phy_buf_trans to increase it's
size.

Nevertheless, the actual value from each Snps's VS/PE-O table is fully
encoded in the lowest byte. Extract the lowest byte by casting to u8 and
using it as deparsed value. This way there is no need to inflate default
tables.

+++ drivers/gpu/drm/i915/display/intel_bios.c
@@ -2202,9 +2202,9 @@ parse_vswing_preemph_snps(union intel_ddi_buf_trans_entry **bufs_mtrx,
 			vals = &tables[offset];
 
 			entry = &bufs_mtrx[idx][row];
-			entry->snps.vswing = vals[0];
-			entry->snps.pre_cursor = vals[1];
-			entry->snps.post_cursor = vals[2];
+			entry->snps.vswing = (u8) vals[0];
+			entry->snps.pre_cursor = (u8) vals[1];
+			entry->snps.post_cursor = (u8) vals[2];
 
 			offset += block->num_columns;
 		}
